chore: Reorganize packages to avoid cyclic imports

+++ view/utils/animation.go
@@ -5,9 +5,9 @@ import (
 	"time"
 )
 
-func AnimateCat(A, B fyne.Position, cat fyne.CanvasObject, iterations float32) {
+func AnimateCat(A, B fyne.Position, cat fyne.CanvasObject, iterations int32) {
 	// TODO: Add scale logic
-	vector := fyne.NewPos((B.X-A.X)/iterations, (B.Y-A.Y)/iterations)
+	vector := fyne.NewPos((B.X-A.X)/float32(iterations), (B.Y-A.Y)/float32(iterations))
 
 	for i := 0; i < int(iterations); i++ {
 		time.Sleep(time.Millisecond)
